jawabAI-backend/requirements.txt
fastapi
uvicorn
python-dotenv
sentence-transformers
langchain-text-splitters
pymupdf
redis
pinecone
google-generativeai
cloudinary
requests
